STM32F207ZG extending PeripheralPins.c to use all available alternate function
Through some minor extensions it is now possible to use all available alternate functions. They exist up to now only as commented lines in PeripheralPins.c.
An API change is not necessary for this new functionality.

The new definitions now looks like:

    {PA_0,            ADC_1, STM_PIN_DATA_EXT(STM_MODE_ANALOG, GPIO_NOPULL, 0, 0, 0)},  // ADC1_IN0
    {PA_0_ALT0,  ADC_2, STM_PIN_DATA_EXT(STM_MODE_ANALOG, GPIO_NOPULL, 0, 0, 0)},  // ADC2_IN0 // choice: PA_0 with ADC_1
    {PA_0_ALT1,  ADC_3, STM_PIN_DATA_EXT(STM_MODE_ANALOG, GPIO_NOPULL, 0, 0, 0)},  // ADC3_IN0 // choice: PA_0 with ADC_1

PA_0, PA_0_ALT0 and PA_0_ALT1 has to be used as pin names for the usage of the three possible ADC blocks connected to the pin (PA_0).